[WEBRTC-732] Move peer related classes into own package.
